requirements.txt
pyqt5-tools==5.15.4.3.2